Use float for moneybox value

+++ app/src/main/java/com/example/minimoneybox/data/MoneyboxData.kt
@@ -1,5 +1,5 @@
 package com.example.minimoneybox.data
 
 data class MoneyboxData(
-    val moneyboxValue: Int
+    val moneyboxValue: Float
 )
